core: reset Turbo model core to upstream; decouple automation
Restore t3.py, s3tokenizer.py, voice_encoder.py to upstream Chatterbox
(removes the repetition-loop guard and the float32 casts). The guard was
firing on natural pauses and chopping audio; upstream generation + the
ASR verification layer are the intended hallucination guard.

- tts_optimized: stop passing guard kwargs to inference_turbo; keep the
  repetition_guard arg as an ignored no-op for CLI back-compat.
- gradio_tts_turbo_app: restore upstream sampling defaults (temp 0.8,
  top_p 0.95, top_k 1000, rep 1.2).
- batch_tts: temperature default 0.5 -> 0.8; verify-asr defaults off.

+++ src/chatterbox/models/s3tokenizer/s3tokenizer.py
@@ -61,7 +61,6 @@ class S3Tokenizer(S3TokenizerV2):
                 wav = torch.from_numpy(wav)
             if wav.dim() == 1:
                 wav = wav.unsqueeze(0)
-            wav = wav.float()  # bugfix ensure float32
 
             n_tokens = (wav.shape[1] / sr) * S3_TOKEN_RATE
             n_tokens = np.ceil(n_tokens)
@@ -84,8 +83,6 @@ class S3Tokenizer(S3TokenizerV2):
                 wav = torch.from_numpy(wav)
             if wav.dim() == 1:
                 wav = wav.unsqueeze(0)
-
-            wav = wav.float()  # bugfix ensure float32
 
             processed_wavs.append(wav)
         return processed_wavs
@@ -117,7 +114,6 @@ class S3Tokenizer(S3TokenizerV2):
             mels.append(mel.squeeze(0))
 
         mels, mel_lens = padding(mels)
-        mels = mels.float()  # bugfix ensure float32 before quantize
         if accelerator is None:
             tokenizer = self
         else:
@@ -164,10 +160,8 @@ class S3Tokenizer(S3TokenizerV2):
         )
         magnitudes = stft[..., :-1].abs()**2
 
-#        mel_spec = self._mel_filters.to(self.device) @ magnitudes
-#torch cuda 124:
-        mel_spec = self._mel_filters.to(self.device, dtype=magnitudes.dtype) @ magnitudes
-        
+        mel_spec = self._mel_filters.to(self.device) @ magnitudes
+
         log_spec = torch.clamp(mel_spec, min=1e-10).log10()
         log_spec = torch.maximum(log_spec, log_spec.max() - 8.0)
         log_spec = (log_spec + 4.0) / 4.0
